fix: use uint32 for timestamp to enable RLP encoding

+++ common/message/message.go
@@ -36,8 +36,8 @@ type AuthMsg struct {
 type Identity struct {
 	// Roller name
 	Name string `json:"name"`
-	// Time of message creation
-	Timestamp int64 `json:"timestamp"`
+	// Unverified Unix timestamp of message creation
+	Timestamp uint32 `json:"timestamp"`
 	// Roller public key
 	PublicKey string `json:"publicKey"`
 	// Version is common.Version+ZkVersion. Use the following to check the latest ZkVersion version.
